Reorganized sidebar component location
